usuarios_service/requirements.txt
flask
flask_sqlalchemy
PyJWT
